remove VLM functionality
