+++ docker-compose.yml
@@ -6,22 +6,3 @@ services:
     container_name: renote_app
     ports:
       - "8501:8501"
-    # environment:
-    #   - OLLAMA_URL=http://ollama:11434
-    volumes:
-      - ./:/app
-    # depends_on:
-    #   - ollama
-
-#   # Опционально: сервис Ollama для VLM режима
-#   ollama:
-#     image: ollama/ollama:latest
-#     container_name: ollama
-#     restart: unless-stopped
-#     ports:
-#       - "11434:11434"
-#     volumes:
-#       - ollama:/root/.ollama
-
-# volumes:
-#   ollama:
