+++ src/renote/processor.py
@@ -4,17 +4,13 @@ from dataclasses import dataclass
 from typing import Optional
 
 from .transforms import split_into_thirds_direct
-from .vlm import classify_slides
+from .scenarios import classify_slides
 
 
 @dataclass
 class ProcessingOptions:
     title_min_font_pt: float = 36.0
     title_min_width_ratio: float = 1.2
-    vlm_mode: str = "heuristic"  # heuristic | ollama
-    soffice_path: Optional[str] = None
-    ollama_model: str = "llava:latest"
-    ollama_url: str = "http://localhost:11434"
 
 
 def process_pptx(
@@ -34,15 +30,10 @@ def process_pptx(
     """
     opts = options or ProcessingOptions()
 
-    # Decide scenario per slide using VLM/heuristics
     scenarios = classify_slides(
         input_pptx,
-        mode=opts.vlm_mode,
         title_min_font_pt=opts.title_min_font_pt,
         title_min_width_ratio=opts.title_min_width_ratio,
-        soffice_path=opts.soffice_path,
-        ollama_model=opts.ollama_model,
-        ollama_url=opts.ollama_url,
     )
 
     split_into_thirds_direct(
