--- ТЗ задачи.md
@@ -1,87 +1,0 @@
-**ТЗ задачи**
-
-​​**1\. Общая информация**
-
-* **Название проекта (рабочее):** Триптихкат  
-* **Цель:** создать интеллектуальный инструмент, который автоматически адаптирует широкоформатные слайды (например, триптихи или диптихи) к стандартному формату 16:9 без потери смысла и визуальной логики, с возможностью дальнейшей печати на А4.  
-* **Целевая аудитория:** пользователи PowerPoint, которым нужно адаптировать презентации под печать или показ.
-
-### **2\. Основной функционал**
-
-#### **2.1. Загрузка исходной презентации**
-
-* Поддержка форматов: .pptx, .pdf,   
-* Автоопределение нестандартных пропорций (например, 32:9, 48:9 и т.п.).
-
-#### **2.2. Семантический разбор слайдов**
-
-* Распознавание блоков, текстов, заголовков, изображений.  
-* Определение логической структуры (колонки, смысловые секции).  
-* Слияние визуального и текстового анализа (визуальные кластеры \+ текстовые метки).
-
-#### **2.3. Интеллектуальная нарезка**
-
-* Алгоритм «умного разрезания» широкого слайда на несколько 16:9 слайдов.  
-* Сохранение заголовков, блоков, линий связей, стрелок.  
-* Автоматическое предложение нарезки \+ возможность ручной корректировки.  
-* Обязательная нумерация презентации.
-
-#### **2.4. Визуальное редактирование**
-
-* Канва предпросмотра: пользователь может передвинуть элементы или вручную объединить/разделить слайды.  
-* Возможность замены элементов (иконок, схем) при переходе в новый формат.  
-* Поддержка адаптации под печать (выровненные отступы, поля A4).
-
-### **3\. ИИ-функции**
-
-* Автоматическое понимание структуры и смысловых блоков.  
-* Рекомендации по нарезке: где «отрезать», как лучше скомпоновать смысл.  
-* Автоматическое восстановление логики: перенос заголовка, повтор ключевой визуализации и т.д.  
-* Оценка сохранения смысла после нарезки (проверка целостности).
-
-### **4\. Визуализация и интерфейс**
-
-#### **4.1. Рабочая зона**
-
-* Просмотр исходного широкоформатного слайда.  
-* Предпросмотр нарезанных слайдов 16:9.  
-* Возможность ручной настройки разреза (drag\&drop, split/join).
-
-#### **4.2. Опции представления**
-
-* Режим “для печати” (A4, без лишнего оформления).  
-* Режим “презентации” (оформление \+ адаптация к стандартному экрану).  
-* Возможность выбрать шаблон стиля (цвет, фон, шрифт).
-
-### **5\. Экспорт и интеграции**
-
-#### **5.1. Экспорт**
-
-* Выгрузка презентации в формате .pptx с адаптированными слайдами.  
-* Выгрузка в PDF для печати (по одному слайду на лист A4, по два — опционально).  
-* Экспорт шаблонов нарезки.
-
-### **6\. Пользовательский опыт (UX/UI)**
-
-* Максимально простой сценарий: загрузка — анализ — проверка — экспорт.  
-* Визуальные подсказки (что будет разрезано, куда уйдут элементы).  
-* Возможность откатить изменения, сравнить «до/после».  
-* Поддержка нескольких презентаций в одном проекте.
-
-### **7\. Техническая архитектура**
-
-* **Фронтенд:** React \+ canvas-based визуализация.  
-* **Бэкенд:** Python (FastAPI) или Node.js.  
-* **ИИ-обработка:**  
-  * OCR и анализ слоев в PowerPoint (python-pptx \+ layoutparser или аналог).  
-  * Модель для распознавания структуры и типовых шаблонов (возможно, LLM \+ Vision).  
-* **Хранение:** PostgreSQL \+ blob-хранилище для файлов.  
-* **Генерация .pptx:** на базе python-pptx или кастомного шаблонного движка.
-
-### **8\. Роли пользователей**
-
-* **Гость:** может загрузить и протестировать одну презентацию.  
-* **Пользователь:** доступ к истории проектов, экспорт в PDF и PPTX.  
-* **Премиум:** доступ к стилям, шаблонам оформления, ручной доработке.  
-* **Админ:** управление шаблонами, аналитика использования, модерация.
-
